Moved PartyPage to container components
